Refactor rocsp-tool
